package discovery renamed to controller in common package
