+++ twister2/api/src/java/edu/iu/dsc/tws/api/task/harp/HarpWorker.java
@@ -47,7 +47,7 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 import edu.iu.dsc.tws.common.config.Config;
-import edu.iu.dsc.tws.common.discovery.IWorkerController;
+import edu.iu.dsc.tws.common.controller.IWorkerController;
 import edu.iu.dsc.tws.common.worker.IPersistentVolume;
 import edu.iu.dsc.tws.common.worker.IVolatileVolume;
 import edu.iu.dsc.tws.common.worker.IWorker;
