+++ twister2/common/src/java/edu/iu/dsc/tws/common/controller/IWorkerController.java
@@ -21,7 +21,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.common.discovery;
+package edu.iu.dsc.tws.common.controller;
 
 import java.util.List;
