+++ twister2/examples/src/java/edu/iu/dsc/tws/examples/internal/rsched/BasicAuroraContainer.java
@@ -12,7 +12,7 @@
 package edu.iu.dsc.tws.examples.internal.rsched;
 
 import edu.iu.dsc.tws.common.config.Config;
-import edu.iu.dsc.tws.common.discovery.IWorkerController;
+import edu.iu.dsc.tws.common.controller.IWorkerController;
 import edu.iu.dsc.tws.common.worker.IPersistentVolume;
 import edu.iu.dsc.tws.common.worker.IVolatileVolume;
 import edu.iu.dsc.tws.common.worker.IWorker;
